좋아요 기능 구현(2022/12/10)

+++ src/main/java/com/cos/photogramstart/web/UserController.java
@@ -2,6 +2,7 @@ package com.cos.photogramstart.web;
 
 import com.cos.photogramstart.config.auth.PrincipalDetails;
 import com.cos.photogramstart.domain.user.User;
+import com.cos.photogramstart.handler.ex.CustomException;
 import com.cos.photogramstart.sevice.UserService;
 import com.cos.photogramstart.web.dto.user.UserProfileDTO;
 import lombok.RequiredArgsConstructor;
@@ -10,6 +11,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
 
 @RequiredArgsConstructor
 @Controller
@@ -26,8 +28,9 @@ public class UserController {
     }
 
     @GetMapping("/user/{id}/update")
-    public String update(@PathVariable int id, @AuthenticationPrincipal PrincipalDetails principalDetails){
+    public String update(@PathVariable Long id, @AuthenticationPrincipal PrincipalDetails principalDetails){
 
         return "user/update";
     }
+
 }
